- Adicionado versão mais completa do gitIgnore - Refatorado POM     - atualizado versãp springboot e java13     - adicionado novas depedências e propriedades     - adicionado processors do lombok e mapstruct para gerar .class durante execução do maven.     - Adicionado validação de formatação coveo fmt-maven - Refatoração do controlador Aquario e suas dependências, como: dto, testes, serviço e repositorio.     - refatorado todos IDs para Long e nomes de tabelas para minusculo     - alterado controle de execeção que gerava um http 404 quando não encontrava algum objeto para retornar http 400.     - removido buscadas por status ativo true, pois aquarios poderão ser removidos.     - removido do banco mariadb para mysql     - adicionado configurações específicas para ambiente de teste, como banco em memória, trace do hibernate e dados inseridos manualmente. - Adicionado MapStruct do aquario para fazer conversão entre objetos, especialmente DTO <-> MODEL - Adicionado serviço para verificar versão da aplicação - Formatado arquivos com coveo ftm-maven

+++ src/main/java/com/paulorpc/aquarium/api/repositories/AquarioRepository.java
@@ -1,20 +1,17 @@
 package com.paulorpc.aquarium.api.repositories;
 
+import com.paulorpc.aquarium.api.entities.Aquario;
 import java.util.List;
 import java.util.Optional;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.paulorpc.aquarium.api.entities.Aquario;
-
 @Repository
-public interface AquarioRepository extends JpaRepository<Aquario, Integer> {
+public interface AquarioRepository extends JpaRepository<Aquario, Long> {
 
   // TODO ALTERAR O PADRÃO DE FIND/FINDALL COM STATUS=TRUE IMPLICITO
 
-  Optional<Aquario> findByIdAndStatusIsTrue(int id);
+  Optional<Aquario> findByIdAndStatusIsTrue(Long id);
 
   List<Aquario> findByStatusIsTrue();
-
 }
