+++ src/main/java/com/paulorpc/aquarium/api/repositories/BiotaRepository.java
@@ -1,16 +1,15 @@
 package com.paulorpc.aquarium.api.repositories;
 
+import com.paulorpc.aquarium.api.entities.Biota;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import com.paulorpc.aquarium.api.entities.Biota;
 
 @Repository
-public interface BiotaRepository extends JpaRepository<Biota, Integer> {
+public interface BiotaRepository extends JpaRepository<Biota, Long> {
 
-  Optional<Biota> findByIdAndDeletadoIsFalse(int id);
+  Optional<Biota> findByIdAndDeletadoIsFalse(Long id);
 
   List<Biota> findByDeletadoIsFalse();
-
 }
